+++ src/cmdliner_base.mli
@@ -24,6 +24,8 @@ val quote : string -> string
 val alts_str : ?quoted:bool -> string list -> string
 val err_ambiguous : kind:string -> string -> ambs:string list -> string
 val err_unknown : ?hints:string list -> kind:string -> string -> string
+val err_multi_def :
+  kind:string -> string -> ('b -> string) -> 'b -> 'b -> string
 
 (** {1:conv Textual OCaml value converters} *)