+++ test/test_term_dups.ml
@@ -1,8 +1,8 @@
 open Cmdliner
 
-let dups p p_dup o o_dup o' =
+let dups p p_dup o o_dup =
   let b = string_of_bool in
-  print_endline (String.concat "\n" [p; p_dup; b o; b o_dup; b o'])
+  print_endline (String.concat "\n" [p; p_dup; b o; b o_dup;])
 
 let test_pos =
   let p =
@@ -13,11 +13,7 @@ let test_pos =
     let doc = "This should show up only once in the docs" in
     Arg.(value & flag & info ["f"; "flag"] ~doc)
   in
-  let o' =
-    let doc = "This is another def for $(opt) so it will show up too" in
-    Arg.(value & flag & info ["flag"] ~doc)
-  in
-  Term.(const dups $ p $ p $ o $ o $ o'),
+  Term.(const dups $ p $ p $ o $ o),
   Term.info "test_term_dups" ~doc:"Test multiple term usage"
 
 let () = match Term.eval test_pos with